migrate: move schema from base to ai
